requirements.txt
Django==2.2.6
wagtail==2.6.3
wagtail-grapple==0.3.1
